# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Production-oriented REST API for the OIG Police Management System.",
        "title": "OIG Police Management API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/members": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns deterministic paginated items and the same page grouped by generation for frontend rendering.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Members"
                ],
                "summary": "List and filter police members",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search Police ID, name, rank, department, or unit",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Employment status",
                        "name": "empFilter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Active penalty count: 0, 1, 2, or 3 for 3+",
                        "name": "discFilter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Department",
                        "name": "deptFilter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Generation or inclusive min-max",
                        "name": "genRangeFilter",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 500,
                        "type": "integer",
                        "default": 100,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.MemberList"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Members"
                ],
                "summary": "Create a police member",
                "parameters": [
                    {
                        "description": "Member",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateMemberRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.Member"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/members/bulk": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Members"
                ],
                "summary": "Atomically create up to 100 members",
                "parameters": [
                    {
                        "description": "Members",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BulkCreateMembersRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.Member"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
...
        "/api/v1/members/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Members"
                ],
                "summary": "Get member profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stable member ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.MemberDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Members"
                ],
                "summary": "Soft-delete a member",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stable member ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Members"
                ],
                "summary": "Update a member and automatically record name history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stable member ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateMemberRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.Member"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
...
        "/api/v1/members/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Members"
                ],
                "summary": "Update employment status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stable member ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateMemberStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.Member"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.BulkCreateMembersRequest": {
            "type": "object",
            "required": [
                "members"
            ],
            "properties": {
                "members": {
                    "type": "array",
                    "maxItems": 100,
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/dto.CreateMemberRequest"
                    }
                }
            }
        },
        "dto.CreateMemberRequest": {
            "type": "object",
            "required": [
                "fullName",
                "generation",
                "policeId"
            ],
            "properties": {
                "department": {
                    "type": "string",
                    "example": "Internal Affairs"
                },
                "employmentStatus": {
                    "type": "string",
                    "example": "รับราชการ"
                },
                "fullName": {
                    "type": "string",
                    "example": "Somchai Jaidee"
                },
                "generation": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 12
                },
                "medicalPrefix": {
                    "type": "string",
                    "example": "Dr."
                },
                "policeId": {
                    "type": "string",
                    "example": "OIG-001"
                },
                "rank": {
                    "type": "string",
                    "example": "Pol. Col."
                },
                "unit": {
                    "type": "string",
                    "example": "Investigation"
                }
            }
        },
...
        "dto.Member": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "employmentStatus": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "generation": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "medicalPrefix": {
                    "type": "string"
                },
                "penaltyCount": {
                    "type": "integer"
                },
                "policeId": {
                    "type": "string"
                },
                "rank": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "dto.MemberDetail": {
            "type": "object",
            "properties": {
                "member": {
                    "$ref": "#/definitions/dto.Member"
                },
                "nameHistory": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.NameHistory"
                    }
                }
            }
        },
        "dto.MemberList": {
            "type": "object",
            "properties": {
                "groupedByGeneration": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/dto.Member"
                        }
                    }
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Member"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dto.Pagination"
                }
            }
        },
...
        "dto.NameHistory": {
            "type": "object",
            "properties": {
                "changedAt": {
                    "type": "string"
                },
                "changedBy": {
                    "type": "string"
                },
                "generation": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "newName": {
                    "type": "string"
                },
                "oldName": {
                    "type": "string"
                },
                "policeId": {
                    "type": "string"
                }
            }
        },
        "dto.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 20
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
...
        "dto.UpdateMemberRequest": {
            "type": "object",
            "properties": {
                "department": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "generation": {
                    "type": "integer",
                    "minimum": 1
                },
                "medicalPrefix": {
                    "type": "string"
                },
                "rank": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateMemberStatusRequest": {
            "type": "object",
            "required": [
                "employmentStatus"
            ],
            "properties": {
                "employmentStatus": {
                    "type": "string",
                    "example": "ปลดราชการ"
                }
            }
        },
...
        "response.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/response.ErrorBody"
                },
                "message": {
                    "type": "string",
                    "example": "ok"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {}
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer {token}\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}